feat: Add data_check.py module.
data_check.py module does a quality check on the tables in two ways. First,
it checks whether if there is a proper amount of rows in each of the coin tables.
Then, it checks if the first date in the table is the same as the first date in the
coin.csv file.